Switch linting to airbnb
XO is not strict enough and AirBnB is the most popular
set of linting/formatting rules. By using the new rules errors
such as using console.log and .only are now detected.

Prettier remains responsible for code formatting as it is slightly
smarter at that, so the eslint config uses settings from prettier
that cause it to ignore Airbnb format rules.