Create a base express app
In time this express will grow to become a
gallery to help developers see their components
and allow visual regression. This first step just
adds a basic express app that can use nunjucks

The Lint rule change is due to the fact that the
gallery, and the dependencies it has are not
shipped as part of the npm package.